src/client/app/phoenix-frontend.component.ts
import {Component} from 'angular2/core';
import {RouteConfig, ROUTER_DIRECTIVES, ROUTER_PROVIDERS} from 'angular2/router';

@Component({
  moduleId: __moduleName,
  selector: 'phoenix-frontend-app',
  providers: [ROUTER_PROVIDERS],
  templateUrl: 'phoenix-frontend.component.html',
  styleUrls: ['phoenix-frontend.component.css'],
  directives: [ROUTER_DIRECTIVES],
  pipes: []
})
@RouteConfig([
])
export class PhoenixFrontendApp {
  defaultMeaning: number = 42;

  meaningOfLife(meaning?: number) {
    return `The meaning of life is ${meaning || this.defaultMeaning}`;
  }
}
